requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.12.0
Pillow==10.0.1
pandas==2.1.1
openpyxl==3.1.2
pytesseract==0.3.10
python-dateutil==2.8.2
requests==2.31.0  # For SW Sapien API calls
pytest==7.4.2
pytest-cov==4.1.0
